// src/lib/schemas.ts
export const LastFmImage = z.object({
  "#text": z.string(),
  size: z.union([z.literal('small'), z.literal('medium'), z.literal('large'), z.literal('extralarge'), z.literal('mega'), z.literal("")]),
})
// ...
export const LastFmTopArtists = z.object({
  streamable: z.string(),
  image: z.array(LastFmImage),
  mbid: z.string(),
  url: z.string().url(),
  playcount: z.string(),
  '@attr': z.object({
    rank: z.string(),
  }),
  name: z.string(),
})